pyproject.toml
[tool.poetry]
name = "cannatracker-back"
version = "0.1.0"
description = ""
authors = ["Thomas <thomas.schleicher@fastad.de>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
bcrypt = "^4.2.0"
fastapi = "^0.112.1"
redis = "^5.0.7"
passlib = "^1.7.4"
pydantic = "^2.8.0"
pyjwt = "^2.8.0"
pytest = "^8.3.2"
python = "^3.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
